refactor: code clean up Move types from App.tsx to types.ts Move excalidrawPackage.test.tsx inside src/tests/package
